requirements.txt
pytest==6.1.0
pytest-html==2.1.1
allure-pytest==2.8.18
PyYAML==5.3.1
selenium==3.141.0
zmail==0.2.8